VueComponentsMainPage: Remove "v-on" explicit directive from event binding, move StartPage into corresponding folder inside of "pages"
